% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caples-package.R
\docType{package}
\name{caples-package}
\alias{caples}
\alias{caples-package}
\title{caples: Fast, Plain-English Readouts for A/B/n Tests}
\description{
Sizes and reads out A/B/n tests on conversion rates in a few function calls. Supports unequal traffic splits, multiple treatment arms with multiplicity correction, sample ratio mismatch checks, achieved minimum detectable effects, and plain-English summaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RobinMahachi/caples}
  \item Report bugs at \url{https://github.com/RobinMahachi/caples/issues}
}

}
\author{
\strong{Maintainer}: Robin Mahachi \email{robinmahachi24@gmail.com}

Authors:
\itemize{
  \item Robin Mahachi \email{robinmahachi24@gmail.com}
}

}
\keyword{internal}
